requirements.txt
aiogram==3.10.0
beautifulsoup4==4.12.3
pdfminer.six==20220524
requests==2.32.3
scikit-learn==1.5.2
numpy==2.1.1
scipy==1.13.1
tqdm==4.66.4
python-dotenv==1.0.1
